finish fix asset system

+++ Source/APILibrary/Win32/Message.cs
@@ -34,7 +34,7 @@ namespace APILibrary.Win32
             => (int)number & 0xffff;
 
         public static int HighWord(IntPtr number)
-            => ((int)number >> 16) & 0xffff;
+            => (short)(((int)number >> 16) & 0xffff);
 
         public static int GetXFromLparam(IntPtr number)
             => (short)((int)number & 0xffff);
